TMSDATAQUALITYCHECKS_AI/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pyspellchecker==0.7.2
gender-guesser==0.4.2
pypostal==1.1
transformers==4.35.0
torch==2.1.0
accelerate==0.24.1
sentencepiece==0.1.99
protobuf==4.25.0

